proyectoprincipal/mainapp/serializers.py
from rest_framework.serializers import ModelSerializer
from django.contrib.auth.models import User


class UserListSerializer(ModelSerializer):
    class Meta:
        model = User
        fields = [
            'id',
            'username',
            'first_name',
            'last_name',
            'email',
            'is_active',
            'is_superuser'
        ]