requirements.txt
Django==2.1.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pillow==5.3.0
djangorestframework==3.9.0